Fixed Multiples bugs including bad cast

+++ Assets/Plugins/Reflexity/MainNodes/UtilityNode.cs
@@ -8,22 +8,22 @@ namespace Plugins.Reflexity.MainNodes {
     public class UtilityNode : MiddleNode {
         
         [Input(ShowBackingValue.Unconnected, ConnectionType.Override), Tooltip("Scale the 0 on the X axe")] 
-        public int MinX;
+        public float MinX;
         [Input(ShowBackingValue.Unconnected, ConnectionType.Override), Tooltip("Evaluated between Min X and Max X")] 
-        public int X;
+        public float X;
         [Input(ShowBackingValue.Unconnected, ConnectionType.Override), Tooltip("Scale the 1 on the X axe")] 
-        public int MaxX = 1;
+        public float MaxX = 1;
         [Tooltip("Evaluate the Utility Y using the X values")]
         public AnimationCurve Function = AnimationCurve.Linear(0, 0, 1, 1);
         [Output(connectionType: ConnectionType.Override), Tooltip("Connect to the Option Node")] public int Rank;
-        public int MaxY = 5;
-        public int MinY = -5;
+        public float MaxY = 5;
+        public float MinY = -5;
         
         public override object GetValue(NodePort port) {
             if (port.fieldName == nameof(Rank)) {
-                int minX = GetInputValue(nameof(MinX), MinX);
-                int maxX = GetInputValue(nameof(MaxX), MaxX);
-                int x = GetInputValue(nameof(X), X);
+                float minX = GetInputValue(nameof(MinX), MinX);
+                float maxX = GetInputValue(nameof(MaxX), MaxX);
+                float x = GetInputValue(nameof(X), X);
                 float scaledX = ScaleX(minX, maxX, x);
                 return ScaleY(MinY, MaxY, Function.Evaluate(scaledX));
             }
@@ -37,7 +37,7 @@ namespace Plugins.Reflexity.MainNodes {
             return (x - minX) / (maxX - minX);
         }
         
-        private int ScaleY(int minY, int maxY, float y) {
+        private int ScaleY(float minY, float maxY, float y) {
             if (Math.Abs(minY - maxY) <= 0) return 0;
             return (int) ((maxY - minY) * y + minY);
         }
